Made index redirect to new fund-specific page, tidied tests
