+++ tests/test_routes.py
@@ -1,7 +1,3 @@
-"""
-Tests the routes and their contents using the dict within
-"route_testing_conf.py".
-"""
 from unittest import mock
 
 import pytest
@@ -11,47 +7,6 @@ from bs4 import BeautifulSoup
 from config import Config
 from flask import render_template
 from requests import HTTPError
-from tests.route_testing_conf import routes_and_test_content
-
-
-def test_routes_status_code(flask_test_client, monkeypatch, mocker):
-    """
-    GIVEN Our Flask Application
-    WHEN a route is requested
-    THEN check that the get response is successful
-    If this test succeeds then our flask application's
-    routes are correctly initialised.
-    """
-
-    monkeypatch.setattr(
-        "fsd_utils.authentication.decorators._check_access_token",
-        lambda: {"accountId": "test-user"},
-    )
-    for route, _ in routes_and_test_content.items():
-        response = flask_test_client.get(route, follow_redirects=True)
-        assert (
-            200 == response.status_code
-        ), f"Incorrect status code returned for route {route}"
-
-
-def test_routes_content(flask_test_client, monkeypatch):
-    """
-    GIVEN Our Flask Application
-    WHEN a route is requested
-    THEN check that the get response contains the
-    expected content.
-
-    If this test succeedes then our flask application's
-    routes are correctly initialised.
-    """
-
-    monkeypatch.setattr(
-        "fsd_utils.authentication.decorators._check_access_token",
-        lambda: {"accountId": "test-user"},
-    )
-    for route, should_contain_this in routes_and_test_content.items():
-        response = flask_test_client.get(route, follow_redirects=True)
-        assert should_contain_this in response.data, f"Error in route {route}"
 
 
 def test_dodgy_url_returns_404(flask_test_client):
@@ -65,17 +20,6 @@ def test_dodgy_url_returns_404(flask_test_client):
     """
     response = flask_test_client.get("/rubbish", follow_redirects=True)
     assert response.status_code == 404
-
-
-def test_page_title_includes_heading(flask_test_client):
-    response = flask_test_client.get("/", follow_redirects=True)
-    soup = BeautifulSoup(response.data, "html.parser")
-    assert (
-        soup.title.string
-        == "Start or continue an application for funding to save an asset in"
-        " your community - Apply for funding to save an asset in your"
-        " community"
-    )
 
 
 def test_page_footer_includes_correct_title_and_link_text(flask_test_client):
@@ -162,3 +106,19 @@ def test_inject_service_name(
         templates_rendered[0][1]["service_title"]
         == "Apply for " + expected_title
     )
+
+
+def test_healthcheck(flask_test_client):
+    response = flask_test_client.get("/healthcheck")
+
+    expected_dict = {
+        "checks": [{"check_flask_running": "OK"}],
+        "version": "abc123",
+    }
+    assert response.status_code == 200, "Unexpected status code"
+    assert response.json == expected_dict, "Unexpected json body"
+
+
+@pytest.mark.app(debug=False)
+def test_app(app):
+    assert not app.debug, "Ensure the app not in debug mode"
