+++ tests/test_start_page.py
@@ -44,6 +44,12 @@ def mock_get_round(mocker):
     )
 
 
+def test_old_index_redirect(client):
+    result = client.get("/", follow_redirects=False)
+    assert result.status_code == 302
+    assert result.location == "/cof/r2w3"
+
+
 def test_start_page_unknown_fund(client, mocker):
     mocker.patch(
         "app.default.routes.get_fund_data_by_short_name", return_value=None
